Revert "Fix size/sign conversion warnings"
This reverts commit 1c5cc3cf27f5c446da8d7d311762765877e26bb5.

+++ packages/seacas/libraries/suplib_cpp/iqsort.C
@@ -73,7 +73,8 @@ namespace {
     }
   }
 
-  template <typename T, typename INT> INT median3(const T v[], INT iv[], size_t left, size_t right)
+  template <typename T, typename INT>
+  size_t median3(const T v[], INT iv[], size_t left, size_t right)
   {
     size_t center = (left + right) / 2;
     size_t pl     = left;
@@ -100,7 +101,7 @@ namespace {
   template <typename T, typename INT> void iqsort(const T v[], INT iv[], size_t left, size_t right)
   {
     if (left + QSORT_CUTOFF <= right) {
-      INT    pivot = median3(v, iv, left, right);
+      size_t pivot = median3(v, iv, left, right);
       size_t i     = left;
       size_t j     = right - 1;
